Include postgres conf file for scram-sha-256 hashes
From now on the postgres playbook will not only deploy a pg_hba.conf
file for determining the authentification methods for the different
users, but also a postgres config file (postgres.conf) for enabling
using password hashed with scram-sha-256. The reason for using a
template to replace the exisiting config file instead of simply
appending to it, is in case further customizations want to be added in
the future.