Change method for locating postgresql config file
Searching for the configuration files by name was problematic, since the
script would simply select the first found, and sometimes there was
duplicates of some config files in caches and other temporary folders,
which lead to the playbyook modifying the wrong "postgresql.conf" file,
and therefore not succeding. Instead the playbook will invoke the psql
cli tool to find the config files.