Make home directory of service user be that of the project
The user created for running respectivly the webserver and database
server, will now both have home directory in the root of the project
directory (so /srv/{project_name}) in contrast to before where the user
simply had no home directory. This is neccessary as some programs will
use the home directory as a place to dump different temporary files like
cache (like firefox using the geckodriver used for scraping) and will
refuse to run without.