Update how generating locale works
Prior the script would simply append the us locale to the locale.gen
file and generate the locales, but this meant that running the script
multiple time would generate the same locale more than one time as the
same locale has then been appended to the file multiple time. Now it
will instead check if the locale is already specified in the file,
append it if it's not and the generate the locales.