Majorly restructure playbooks
First of all, all variables will be set based on those from the JSON
files in the var directory. Secondly all playbooks - except for the
main.yml in the playbooks directory - will only contain tasks, as
instead of importing playbooks, they will be including tasks. The reason
for this is to not have to pass down variable manually through multiple
layers of playbooks, but instead simply simply import all of them from
the JSON files in the main.yml playbooks in the root playbook dir, and
then possibly creating more variables based on those using set_fact
later down the chain of playbooks. Both debian and arch is officially
supported now.